requirements.txt
# Créer le fichier
python-dotenv
typing
jsonpath_ng
google-cloud-aiplatform
neo4j
graphdatascience
mcp
mcp-neo4j-cypher
requests
aiohttp
grpcio
protobuf
rich
typing-extensions
netmiko